moved ProcessorData to a separate class
